Do not use python-config
python-config is not included in most virtualenv, hence we should not
use it. For now only detect PYTHON_INCLUDES, more can be added later as
needed. Now it SHOULD use the proper include path set for the Pythn
interpreter set with --with-python, whether it is a Python installation
or only a virtual environment.

In case of virtual environments, both the script and extension module
directories are set to ${prefix}/... or ${exec_prefix}/..., hence it is
required to specify --prefix=/path/to/virtualenv on configure to install
in virtual environment.

Do we want to change this for virtual environments to by default point
to the virtual environment's site-package installation locations?