Do not add cgpt/lib to include path
Depending on the order the compiler looks for files in include paths
this causes issues as headers in cgpt/lib may overwrite global headers.
In this particular case cgpt/lib/time.h was included instead of
/usr/include/time.h (for get_clocktime).
This should not be an issue as no cgpt headers are included via "<>".
Was actually required to find the non-used config.h. With the changes in
mean-time this file is know in topdir - this directory is still added as
include path.